Monthely Payment uploading with utility and Junit

+++ main/dayOfWeek.java
@@ -5,7 +5,7 @@ import main.Utility.UtilityForDayOfWeek;
 public class dayOfWeek {
 	public static void main(String[] args) {
 		
-		int y=0,m=0,d=0;
+		double y=0,m=0,d=0;
 		
 		d = Integer. parseInt(args[0]);
 		m = Integer. parseInt(args[1]);
